fix: Add Flyway migration V7 to correct macro preset column types, include a repair script for V5 checksum, and provide production application configuration and deployment documentation.

+++ src/main/resources/db/migration/V5__macro_presets.sql
@@ -3,16 +3,15 @@ CREATE TABLE macro_presets (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     user_id BIGINT NOT NULL,
     name VARCHAR(100) NOT NULL,
-    protein_percentage INT NOT NULL,  -- 0-100
-    carbs_percentage INT NOT NULL,    -- 0-100
-    fats_percentage INT NOT NULL,     -- 0-100
+    protein_percentage DECIMAL(5, 2) NOT NULL,  -- 0-100
+    carbs_percentage DECIMAL(5, 2) NOT NULL,    -- 0-100
+    fats_percentage DECIMAL(5, 2) NOT NULL,     -- 0-100
     is_default BOOLEAN DEFAULT FALSE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-    
+    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, 
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE,
     INDEX idx_user_id (user_id),
-    
+
     -- Ensure sum of percentages is 100 (checked in application layer)
     CONSTRAINT chk_percentages CHECK (
         protein_percentage >= 0 AND protein_percentage <= 100 AND
